autoapply/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
google-generativeai==0.8.0
httpx==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
lxml==5.3.0
google-api-python-client==2.118.0
google-auth==2.28.0
playwright==1.47.0
